Added some helper functions
Added helper functions for getting the previous kill, the XY position of a kill given an index, and also calculating the distance of that kill from the blue teams' nexus. Also added helper functions for getting the last structure taken, as well as a helper function to return a "game state" given a minute.